实现报单先撤销再删除

+++ report-react/src/containers/Body.js
@@ -6,7 +6,7 @@ import Prompt from '../components/Prompt'
 import BillDetail from '../containers/BillDetail';
 import DeviceDetail from '../containers/DeviceDetail';
 import { connect } from 'react-redux';
-import { addBillItem, deleteBillItem } from '../reducers/reducer';
+import { addBillItem, deleteBillItem, cancelBillItem } from '../reducers/reducer';
 import PropTypes from 'prop-types';
 
 class Body extends Component {
@@ -16,13 +16,15 @@ class Body extends Component {
         billList: PropTypes.arrayOf(PropTypes.object),
         deviceList: PropTypes.arrayOf(PropTypes.object),
         onAddBillItem: PropTypes.func,
-        onDeleteBillItem: PropTypes.func
+        onDeleteBillItem: PropTypes.func,
+        onCancelBillItem: PropTypes.func
     }
 
     constructor() {
         super();
         this.state = {
             index: -1,
+            type: '',
             takeBill: false,
             showPrompt: false,
             showDetail: false,
@@ -44,6 +46,14 @@ class Body extends Component {
         localStorage.setItem('report_data', JSON.stringify(data));
     }
 
+    //在localstorage中取消一条报单
+    _cancelBillItem(index) {
+        const data = JSON.parse(localStorage.report_data);
+        data.billList[index].billStatus = '已取消';
+        localStorage.setItem('report_data', JSON.stringify(data));
+    }
+
+
     //处理报单的提交
     handleOnSubmit(billItem) {
         this._addBillItem(billItem);
@@ -59,6 +69,17 @@ class Body extends Component {
     handleOnCancel(index) {
         this.setState({
             index: index,
+            type: 'cancel',
+            showDetail: false,
+            showPrompt: true
+        });
+    }
+
+    //在列表项点击删除
+    handleOnDelete(index) {
+        this.setState({
+            index: index,
+            type: 'delete',
             showDetail: false,
             showPrompt: true
         });
@@ -74,14 +95,12 @@ class Body extends Component {
 
     //在详细信息中点击再次/马上报修
     handleOnBill(e,load) {
-
         this.setState({
             takeBill: true,
             showDetail: false,
             //load是为了实现再次/立即报修时将信息load进表单中
             load: load
         });
-        console.log(this.state.load)
     }
 
     //点击遮布或者关闭按钮
@@ -95,11 +114,18 @@ class Body extends Component {
         });
     }
 
-     //确认撤销
-     handleOnConfirmCancel() {
-        this._deleteBillItem(this.state.index);
-        if (this.props.onDeleteBillItem) {   
-            this.props.onDeleteBillItem(this.state.index);
+     //在提示中点击确认
+     handleOnConfirm() {
+        if (this.state.type === 'delete') {
+            this._deleteBillItem(this.state.index);
+            if (this.props.onDeleteBillItem) {   
+                this.props.onDeleteBillItem(this.state.index);
+            }
+        } else if (this.state.type === 'cancel') {
+            this._cancelBillItem(this.state.index);
+            if (this.props.onCancelBillItem) {   
+                this.props.onCancelBillItem(this.state.index);
+            }
         }
         this.setState({
             showPrompt: false,
@@ -107,11 +133,12 @@ class Body extends Component {
         });
     }
 
-     //取消撤销
-     handleOnRecallCancel() {
+     //在提示中点击取消
+     handleOnRecall() {
         this.setState({
-            showPrompt: false,
-            index: -1
+            index: -1,
+            type: '',
+            showPrompt: false
         });
     }
 
@@ -121,12 +148,13 @@ class Body extends Component {
                 return(
                     <div className="body">
                         <div id="bill-page">
-                            <List list={this.props.billList} onCancel={this.handleOnCancel.bind(this)} onCheck={this.handleOnCheck.bind(this)} />
+                            <List list={this.props.billList} onCancel={this.handleOnCancel.bind(this)} 
+                                onCheck={this.handleOnCheck.bind(this)} onDelete={this.handleOnDelete.bind(this)} />
                             <BillDetail show={this.state.showDetail} index={this.state.index} list={this.props.billList}
                                 onClose={this.handleOnClose.bind(this)} onCancel={this.handleOnCancel.bind(this)} onBill={this.handleOnBill.bind(this)}
                             />
-                            <Prompt show={this.state.showPrompt} onConfirmClick={this.handleOnConfirmCancel.bind(this)} 
-                                onRecallClick={this.handleOnRecallCancel.bind(this)} onClose={this.handleOnClose.bind(this)}   
+                            <Prompt show={this.state.showPrompt} onConfirmClick={this.handleOnConfirm.bind(this)} 
+                                onRecallClick={this.handleOnRecall.bind(this)} onClose={this.handleOnClose.bind(this)}   
                             />
                         </div>
                         <TakeBill show={this.state.takeBill} page={this.props.page} load={this.state.load}
@@ -138,7 +166,7 @@ class Body extends Component {
                 return(
                     <div className="body">
                         <div id="device-page">
-                            <List list={this.props.deviceList} onCheck={this.handleOnCheck.bind(this)} />
+                            <List list={this.props.deviceList} onCheck={this.handleOnCheck.bind(this)} onDelete={this.handleOnDelete.bind(this)} />
                             <DeviceDetail show={this.state.showDetail} index={this.state.index} list={this.props.deviceList} 
                                 onClose={this.handleOnClose.bind(this)} onBill={this.handleOnBill.bind(this)}
                             />
@@ -180,6 +208,9 @@ const mapDispatchToProps = (dispatch) => {
         },
         onDeleteBillItem: (index) => {
             dispatch(deleteBillItem(index));
+        },
+        onCancelBillItem: (index) => {
+            dispatch(cancelBillItem(index));
         }
     }
 }
